requirements.txt
astroid==1.5.3
autopep8==1.3.2
ffmpeg-python==0.1.6
ffmpy==0.2.2
flake8==3.4.1
future==0.16.0
isort==4.2.15
jedi==0.11.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.13.3
opencv-python==3.3.0.10
parso==0.1.0
pep8==1.7.0
pycodestyle==2.3.1
pyflakes==1.5.0
pylint==1.7.4
pyserial==3.4
six==1.11.0
wrapt==1.10.11
